Better Pager implementation
We don't want to be calling into the Pager more than once, and it turns
out that everything we need can be provided by the view.

Since we only care about the new page number, we'll send our own custom
type into the Pager and let it hand us the new page number. This is done
instead of sending msg's from the runtime into the Pager module.

Note that the Pager is stateless, we want to handle that ourselves in
the Main view.

In addition, pull out styles into a stylesheet so they can be overridden,
if desired.